+++ kernel/kernel-test/src/main/java/org/sakaiproject/test/ModiInitializer.java
@@ -1,4 +1,4 @@
-package org.sakaiproject.announcement.tool;
+package org.sakaiproject.test;
 
 import org.sakaiproject.modi.SharedApplicationContext;
 import org.springframework.context.ApplicationContextInitializer;
@@ -9,7 +9,7 @@ import java.nio.file.Path;
 public class ModiInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
     @Override
     public void initialize(ConfigurableApplicationContext applicationContext) {
-        System.setProperty("sakai.home", Path.of("target/sakai-home").toAbsolutePath().toString());
+        System.setProperty("sakai.home", Path.of("target/sakai-home").toAbsolutePath().toString() + "/");
         System.setProperty("sakai.modi.enabled", "true");
     }
 }
