+++ kernel/kernel-test/src/main/java/org/sakaiproject/test/ModiWebTest.java
@@ -1,4 +1,4 @@
-package org.sakaiproject.announcement.tool;
+package org.sakaiproject.test;
 
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.ContextHierarchy;
