+++ kernel/pom.xml
@@ -30,6 +30,7 @@
         <module>kernel-util</module>
         <module>kernel-storage-util</module>
         <module>kernel-test</module>
+        <module>test-helper</module>
     </modules>
 
     <profiles>
