+++ kernel/test-helper/src/main/resources/log4j.properties
@@ -7,9 +7,9 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d{ABSOLUTE} %5p %c{1}:%L - %m%n
 
 ### Sakai ###
-log4j.logger.org.sakaiproject=info
-log4j.logger.org.sakaiproject.component=info
-log4j.logger.org.sakaiproject.test=info
+#log4j.logger.org.sakaiproject=info
+#log4j.logger.org.sakaiproject.component=debug
+#log4j.logger.org.sakaiproject.test=debug
 
 ### Spring and Hibernate ###
 #log4j.logger.org.springframework=debug
